Run successful jobs first
