+++ CRM/Core/JobManager.php
@@ -9,6 +9,8 @@
  +--------------------------------------------------------------------+
  */
 
+use Civi\Api4\Job;
+
 /**
  * This interface defines methods that need to be implemented
  * by every scheduled job (cron task) in CiviCRM.
@@ -24,6 +26,7 @@ class CRM_Core_JobManager {
    * Format is ($id => CRM_Core_ScheduledJob).
    *
    * @var CRM_Core_ScheduledJob[]
+   * @deprecated
    */
   public $jobs = NULL;
 
@@ -47,13 +50,6 @@ class CRM_Core_JobManager {
   public $_source = NULL;
 
   /**
-   * Class constructor.
-   */
-  public function __construct() {
-    $this->jobs = $this->getJobs();
-  }
-
-  /**
    * @param bool $auth
    */
   public function execute($auth = TRUE) {
@@ -67,13 +63,41 @@ class CRM_Core_JobManager {
 
     // it's not asynchronous at this stage
     CRM_Utils_Hook::cron($this);
-    foreach ($this->jobs as $job) {
-      if ($job->is_active) {
-        if ($job->needsRunning()) {
-          $this->executeJob($job);
-        }
+
+    // Get a list of the jobs that have completed previously
+    $successfulJobs = Job::get(FALSE)
+      ->addWhere('is_active', '=', TRUE)
+      ->addClause('OR', ['last_run', 'IS NULL'], ['last_run', '<=', 'last_run_end', TRUE])
+      ->addOrderBy('name', 'ASC')
+      ->execute()
+      ->indexBy('id')
+      ->getArrayCopy();
+
+    // Get a list of jobs that have not completed previously.
+    // This could be because they are a new job that has not yet run or a job that is fatally crashing (eg. OOM).
+    // If last_run is NULL the job has never run and will be selected above so exclude it here
+    // If last_run_end is NULL the job has never completed successfully.
+    // If last_run_end is < last_run job has completed successfully in the past but is now failing to complete.
+    $maybeUnsuccessfulJobs = Job::get(FALSE)
+      ->addWhere('is_active', '=', TRUE)
+      ->addWhere('last_run', 'IS NOT NULL')
+      ->addClause('OR', ['last_run_end', 'IS NULL'], ['last_run', '>', 'last_run_end', TRUE])
+      ->addOrderBy('name', 'ASC')
+      ->execute()
+      ->indexBy('id')
+      ->getArrayCopy();
+
+    $jobs = array_merge($successfulJobs, $maybeUnsuccessfulJobs);
+    foreach ($jobs as $job) {
+      $temp = ['class' => NULL, 'parameters' => NULL, 'last_run' => NULL];
+      $scheduledJobParams = array_merge($job, $temp);
+      $jobDAO = new CRM_Core_ScheduledJob($scheduledJobParams);
+
+      if ($jobDAO->needsRunning()) {
+        $this->executeJob($jobDAO);
       }
     }
+
     $this->logEntry('Finishing scheduled jobs execution.');
 
     // Set last cron date for the status check
@@ -144,27 +168,6 @@ class CRM_Core_JobManager {
 
     // Save the job last run end date (if this doesn't get written we know the job crashed and was not caught (eg. OOM).
     $job->saveLastRunEnd();
-  }
-
-  /**
-   * Retrieves the list of jobs from the database,
-   * populates class param.
-   *
-   * @return array
-   *   ($id => CRM_Core_ScheduledJob)
-   */
-  private function getJobs(): array {
-    $jobs = [];
-    $dao = new CRM_Core_DAO_Job();
-    $dao->orderBy('name');
-    $dao->domain_id = CRM_Core_Config::domainID();
-    $dao->find();
-    while ($dao->fetch()) {
-      $temp = ['class' => NULL, 'parameters' => NULL, 'last_run' => NULL];
-      CRM_Core_DAO::storeValues($dao, $temp);
-      $jobs[$dao->id] = new CRM_Core_ScheduledJob($temp);
-    }
-    return $jobs;
   }
 
   /**
